+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.23.1
+
+* Fix running paths by absolute path (with drive letter) on windows.
+
 ## 1.23.0
 
 * Avoid empty expandable groups for tests without extra output in Github